refactor(compile): split stdlib stub-symbol localization out of strip_dedup.rs
strip_dedup.rs crossed the 2000-line size gate (2087) once this PR's UI-lib
dedup landed. Move the stdlib stub-symbol localization cluster into
strip_dedup/stub_symbols.rs; strip_dedup.rs is now 1572.
